Refactoring

+++ admin/producto_listar.php
@@ -1,10 +1,10 @@
-<? include_once('inc/header.php');?>
+<?php include_once('inc/header.php');?>
 
   <body class="nav-md">
     <div class="container body">
       <div class="main_container">
         <div class="col-md-3 left_col">
-          <? include_once('inc/left_menu.php');?>
+          <?php include_once('inc/left_menu.php');?>
 
         <!-- page content -->
         <div class="right_col" role="main">
@@ -40,19 +40,20 @@
 							$sql = 'SELECT id_producto, Nombre_de_producto, Precio_de_producto, descripcion
                             FROM productos ORDER BY id_producto ASC';
 							$resultado = $con->query($sql);
+              if ($resultado):
 							foreach($resultado as $row){
 								?>
 								  <tr class="even pointer"> 
-									<td class=" "><? echo $row['id_producto']?></td>
+									<td class=" "><?php echo $row['id_producto']?></td>
 									<td class=" "><?= $row['Nombre_de_producto']?> </td> <!--PRODUCTOS-->
-									<td class=" "><? print($row['Precio_de_producto'])?></td> <!--PRECIO DE PRODUCTOS-->
-									<td class=" "><? print($row['descripcion'])?></td><!--DESCRIPCION-->
+									<td class=" "><?php print($row['Precio_de_producto'])?></td> <!--PRECIO DE PRODUCTOS-->
+									<td class=" "><?php print($row['descripcion'])?></td><!--DESCRIPCION-->
 
 										<td class=" last"><a href="productos_editar.php?id=<?=$row['id_producto']?>">Edit</a>
 										<td class=" last"><a href="productos_eliminar.php?id=<?=$row['id_producto']?>">Eliminar</a>
 									</td>
 								  </tr>
-							<?}?>
+							<?} endif;?>
                          
                         </tbody>
                       </table>
@@ -64,4 +65,4 @@
               </div>
         </div>
         <!-- /page content -->
-<? include_once('inc/footer.php');
+<?php include_once('inc/footer.php');
